V57: Depth First search
Depth-First Search (DFS) is a traversal algorithm that explores as far as possible along a branch before backtracking. In a tree, DFS can be implemented in multiple ways (recursive or iterative).

This is the simplest way to implement DFS. The function calls itself for each child node until all nodes are visited.